# Scatter optimization : Eliminate 64b integer divide.