@vj-krish 2024-02-10
## Perf on M1 Max MBP (32-core GPU, 64GB RAM)
### Before

====================
X (1000000, 64), Indices (1000000,)
MLX: 30.969ms
PyTorch: 12.942ms
====================
X (1000000, 64), Indices (1000000,)
MLX: 31.131ms
PyTorch: 13.768ms
====================
X (100000, 64), Indices (100000,)
MLX: 4.480ms
PyTorch: 1.490ms

### After

====================
X (1000000, 64), Indices (1000000,)
MLX: 6.676ms
PyTorch: 12.998ms
====================
X (1000000, 64), Indices (1000000,)
MLX: 6.736ms
PyTorch: 13.778ms
====================
X (100000, 64), Indices (100000,)
MLX: 2.350ms
PyTorch: 1.517ms