@vj-krish 2024-02-10
There was some hypothesis that this kernel is bound by atomics. But this wasn't the case. On profiling, we identified a couple of hotspots  -
1. 64bit integer divide operation was pretty expensive and eliminating only that resulted in all the perf gains associated with this change. Gather kernel's performance also benefits significantly from the same change.
2. The workload is still integer arithmetic bound. I tried moving some ops for buffer offset calculation to fp ops and it helped improve perf by another 25-50% (depending on the shape of inputs). But there are few quirks that we need to think through. I'll try and push this as a separate change along with a few other ideas @awni had in mind.