@TristanBilot 2024-02-10
Thanks for taking the lead on this issue! Interesting findings regarding the expensive 64bit integer division.
I tried to run your code locally but get this error when building:
```
mlx/mlx/backend/metal/make_compiled_preamble.sh: line 15: $OUTPUT_FILE: ambiguous redirect
make[2]: *** [mlx/backend/metal/compiled_preamble.cpp] Error 1
make[1]: *** [mlx/backend/metal/CMakeFiles/compiled_preamble.dir/all] Error 2
make[1]: *** Waiting for unfinished jobs....
```