@vj-krish 2024-02-10
@TristanBilot thanks for sharing benchmarks with mlx-graphs. Results for the scatter benchmarks - 

### Perf on M1 Max MBP (32-core GPU, 64GB RAM)

#### Before

Detailed benchmark:
| Operation                                                      | mlx_gpu | mlx_cpu | pyg_mps | pyg_cpu | mlx_gpu/pyg_cpu speedup | mlx_gpu/pyg_mps speedup |
|----------------------------------------------------------------|-------|-------|-------|-------|-----------------------|-----------------------|
| benchmark_scatter / edg=(2, 100000) nod=(100, 64) sca=add  |   4.92 |  22.79 |  81.71 |   6.28 |    +27% |  +1561% |
| benchmark_scatter / edg=(2, 1000000) nod=(100, 64) sca=add |  31.59 | 227.38 | 875.71 |  72.18 |   +128% |  +2672% |
| benchmark_scatter / edg=(2, 10000) nod=(1000, 64) sca=add  |   1.47 |   5.52 |   2.02 |   1.52 |     +3% |    +37% |
| benchmark_scatter / edg=(2, 100000) nod=(100, 64) sca=max  |   6.33 |  22.78 | nan |  16.89 |   +166% | nan |
| benchmark_scatter / edg=(2, 1000000) nod=(10, 64) sca=max  |  31.52 | 226.82 | nan | 179.21 |   +468% | nan |
| benchmark_scatter / edg=(2, 10000) nod=(1000, 64) sca=max  |   1.79 |   6.52 | nan |  14.78 |   +724% | nan |

 Average benchmark:
| Operation              | mlx_gpu | mlx_cpu | pyg_mps | pyg_cpu | mlx_gpu/pyg_cpu speedup | mlx_gpu/pyg_mps speedup |
|------------------------|-------|-------|-------|-------|-----------------------|-----------------------|
| benchmark_scatter  |  12.94 |  85.30 | nan |  48.48 |   +274% | nan |

#### After

Detailed benchmark:
| Operation                                                      | mlx_gpu | mlx_cpu | pyg_mps | pyg_cpu | mlx_gpu/pyg_cpu speedup | mlx_gpu/pyg_mps speedup |
|----------------------------------------------------------------|-------|-------|-------|-------|-----------------------|-----------------------|
| benchmark_scatter / edg=(2, 100000) nod=(100, 64) sca=add  |   2.72 |  23.77 |  82.78 |   5.72 |   +110% |  +2943% |
| benchmark_scatter / edg=(2, 1000000) nod=(100, 64) sca=add |   7.77 | 226.85 | 986.48 |  73.54 |   +845% | +12588% |
| benchmark_scatter / edg=(2, 10000) nod=(1000, 64) sca=add  |   1.79 |   8.35 |   2.11 |   2.71 |    +50% |    +17% |
| benchmark_scatter / edg=(2, 100000) nod=(100, 64) sca=max  |   2.75 |  23.15 | nan |  17.08 |   +520% | nan |
| benchmark_scatter / edg=(2, 1000000) nod=(10, 64) sca=max  |   7.08 | 229.33 | nan | 177.99 |  +2415% | nan |
| benchmark_scatter / edg=(2, 10000) nod=(1000, 64) sca=max  |   1.40 |   3.26 | nan |  11.43 |   +717% | nan |

 Average benchmark:
| Operation              | mlx_gpu | mlx_cpu | pyg_mps | pyg_cpu | mlx_gpu/pyg_cpu speedup | mlx_gpu/pyg_mps speedup |
|------------------------|-------|-------|-------|-------|-----------------------|-----------------------|
| benchmark_scatter  |   3.92 |  85.79 | nan |  48.08 |  +1126% | nan |