@TristanBilot 2024-02-10
Wow, your results seem very impressive for such a small change!!
It seems @francescofarina also failed to build the branch locally.