@awni 2024-02-10
@TristanBilot could you double check building main works without this PR? The issue sounds related to the compile preamble.