@awni 2024-02-10
@vj-krish this is incredible, such a huge speedup!!

Btw @angeloskath and @jagrit06 turns out a lot of the speedup in the gather that I recently pushed (#626) was actually due to this same reason - using the 2D grid avoids the 64-bit integer division and mod which are evidently EXTREMELY slow. We should watch out for that in any other kernels.